feat: major enhancements for reliability and observability
- Add failover client with RPC failover support
- Implement latency ring for observability metrics
- Add new observability providers and tests
- Enhance reliability with horizontal scaling improvements
- Add admin API key handling and auth improvements
- Implement finality checker for blockchain events
- Add comprehensive testing and benchmarking
- Improve multi-chain indexing capabilities
- Add WebSocket handler tests and validation
- Enhance GRPC authentication and interceptors
- Add Solana puller support and reorg handling
- Improve database adapters and query optimization

+++ pkg/application/indexing/runtime.go
@@ -16,7 +16,7 @@ type EventEnvelope struct {
 	ChainID          string
 	BlockNumber      uint64
 	TransactionHash  string
-	LogIndex         uint
+	LogIndex         uint64
 	Payload          interface{}
 	ReceivedAt       time.Time
 	CheckpointCursor string
